bugfix: 修复获取项目/业务的开发商ID错误的问题 #1932
