+++ pipeline_plugins/base/utils/inject.py
@@ -43,6 +43,27 @@ def supplier_account_for_business(biz_cc_id):
     return supplier_account
 
 
+def supplier_id_for_project(project_id):
+    try:
+        proj = Project.objects.get(id=project_id)
+    except Project.DoesNotExist:
+        return 0
+
+    if not proj.from_cmdb:
+        return 0
+
+    return supplier_id_for_business(proj.bk_biz_id)
+
+
+def supplier_id_for_business(biz_cc_id):
+    try:
+        supplier_id = Business.objects.supplier_id_for_business(biz_cc_id)
+    except Business.DoesNotExist:
+        supplier_id = 0
+
+    return supplier_id
+
+
 def supplier_account_inject(func):
     def wrapper(*args, **kwargs):
         if 'project_id' in kwargs:
@@ -60,11 +81,11 @@ def supplier_account_inject(func):
 def supplier_id_inject(func):
     def wrapper(*args, **kwargs):
         if 'project_id' in kwargs:
-            kwargs['supplier_id'] = supplier_account_for_project(kwargs['project_id'])
+            kwargs['supplier_id'] = supplier_id_for_project(kwargs['project_id'])
         elif 'biz_cc_id' in kwargs:
-            kwargs['supplier_id'] = supplier_account_for_business(kwargs['biz_cc_id'])
+            kwargs['supplier_id'] = supplier_id_for_business(kwargs['biz_cc_id'])
         elif 'bk_biz_id' in kwargs:
-            kwargs['supplier_id'] = supplier_account_for_business(kwargs['bk_biz_id'])
+            kwargs['supplier_id'] = supplier_id_for_business(kwargs['bk_biz_id'])
 
         return func(*args, **kwargs)
