+++ pipeline_plugins/components/utils/common.py
@@ -17,78 +17,10 @@ import os
 
 from django.conf import settings
 
-from gcloud.core.models import Project, Business
-
 logger = logging.getLogger('root')
 
 ip_re = r'((25[0-5]|2[0-4]\d|[01]?\d\d?)\.){3}(25[0-5]|2[0-4]\d|[01]?\d\d?)'
 ip_pattern = re.compile(ip_re)
-
-
-def supplier_account_for_project(project_id):
-    try:
-        proj = Project.objects.get(id=project_id)
-    except Project.DoesNotExist:
-        return 0
-
-    if not proj.from_cmdb:
-        return 0
-
-    return supplier_account_for_business(proj.bk_biz_id)
-
-
-def supplier_id_for_project(project_id):
-    try:
-        proj = Project.objects.get(id=project_id)
-    except Project.DoesNotExist:
-        return 0
-
-    if not proj.from_cmdb:
-        return 0
-
-    return supplier_id_for_business(proj.bk_biz_id)
-
-
-def supplier_account_for_business(biz_cc_id):
-    try:
-        supplier_account = Business.objects.supplier_account_for_business(biz_cc_id)
-    except Business.DoesNotExist:
-        supplier_account = 0
-
-    return supplier_account
-
-
-def supplier_id_for_business(biz_cc_id):
-    try:
-        supplier_id = Business.objects.supplier_id_for_business(biz_cc_id)
-    except Business.DoesNotExist:
-        supplier_id = 0
-
-    return supplier_id
-
-
-def supplier_account_inject(func):
-    def wrapper(*args, **kwargs):
-        if 'project_id' in kwargs:
-            kwargs['supplier_account'] = supplier_account_for_project(kwargs['project_id'])
-        elif 'biz_cc_id' in kwargs:
-            kwargs['supplier_account'] = supplier_account_for_business(kwargs['biz_cc_id'])
-
-        return func(*args, **kwargs)
-
-    return wrapper
-
-
-def supplier_id_inject(func):
-    def wrapper(*args, **kwargs):
-        if 'project_id' in kwargs:
-            kwargs['supplier_id'] = supplier_id_for_project(kwargs['project_id'])
-        elif 'biz_cc_id' in kwargs:
-            kwargs['supplier_id'] = supplier_id_for_business(kwargs['biz_cc_id'])
-
-        return func(*args, **kwargs)
-
-    return wrapper
 
 
 def get_ip_by_regex(ip_str):
